migration part of: cinder: Add configuration option for multipath (bsc#997607)
This patch only includes the changes in the schema.

Currently, mulitpath support is set by default in the cinder and
nova cookbooks, and both cookbooks force it to be true, but the
built package overrides nova's setting to false.

Remove that default support, and add a new option to cinder which allows
operators to enable or disable multipath support for both cinder
and nova.

Cherry picked from ab5df7ef6df1d84a33bd41f4473dd58fdd504542.