neutron: Avoid conflict between br-fixed and br-public (bsc#1040396)
With ovs, when the vlan type driver is used, we usually have a br-fixed
network to deal with vlan networks.

However, when the floating network is not using vlan and is on the same
conduit as the fixed network, then the underlying interface for br-fixed
and br-public is going to be the same. And of course it can't be in both
bridges. This results in the interface moving from one to the other
during a chef run, creating some network outage.

The solution is to simply merge br-fixed in br-public for that case.
This will result in physnet1:br-public as bridge mapping.

The cookbook change is similar to the change that was done for
linuxbridge in https://github.com/crowbar/crowbar-openstack/pull/128

https://bugzilla.suse.com/show_bug.cgi?id=1040396