heat: move keystone-dependant resource to execute phase
We were doing calls to keystone during the compile phase
to obtain the domain id for heat. This can cause issues
when changing the keystone endpoint protocol.

This patch moves the gathering of the domain id to the
execution phase and makes the variables for the heat
config template to be lazy, so the execution is delayed
until the template is written.